refactor: exception code object, object constructor
- 에러 코드 객체명 변경
- 생성자 주입을 명시적으로 선언하도록 변경
