+++ see-realview/src/main/java/com/see/realview/user/controller/UserController.java
@@ -1,9 +1,9 @@
 package com.see.realview.user.controller;
 
 import com.see.realview.core.response.Response;
+import com.see.realview.token.entity.TokenPair;
 import com.see.realview.user.dto.request.LoginRequest;
 import com.see.realview.user.dto.request.RegisterRequest;
-import com.see.realview.user.dto.response.TokenPair;
 import com.see.realview.user.service.UserServiceImpl;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
