+++ see-realview/src/main/java/com/see/realview/user/service/UserService.java
@@ -1,8 +1,8 @@
 package com.see.realview.user.service;
 
+import com.see.realview.token.entity.TokenPair;
 import com.see.realview.user.dto.request.LoginRequest;
 import com.see.realview.user.dto.request.RegisterRequest;
-import com.see.realview.user.dto.response.TokenPair;
 
 public interface UserService {
     void register(RegisterRequest request);
